Convert Words to Morsecode
Today is day 2 of 200 Days of Code. I made this code because I recently had a lesson about text procession in Python. This program was really fun to make as it wasn't complicated and only took me a short amount of time. I have also recently been interested in hidden messages and the ways to communicate these, hence why I chose to do a 'words to morse code' translator program.