Day 3: Clear the Circles
This code is another one that makes use of Tkinter. The 'make_canvas' function was honestly a copy-paste from my Python coding class as I am still a beginner. I just made this code so that I can experiment with the different 'mouse_pressed' commands and this one deletes whatever you press. I have also just been experimenting with shapes in general, and for this program, I made use of random sized circles.