Day 4: Heads Up
I wanted to make a simple game which a lot of people could play. This game called 'Heads Up' wherein one player wouldn't be allowed to look at the screen. S/he gets a word and the other players will have to tell clues so that the player can guess the given word. This is a really fun code to do for a beginner like me!