Day 5: Addition Practice
This was one of my assignments for my coding class. I had so much fun doing this as a beginner. 'AdditionPractice.py' is a program that lets users learn. This program randomly generates addition problems for the user and checks to see if the user got it right or wrong, until the s/he gets 5 correct in a row.