Day 7: Random Name Generator
This program randomly rearranges names the user inputs. My teachers would always use this to assign the order of presentations in my class. After coding a random number generator, I wanted to step it up a bit so I made this program. I recently also learned about lists and how it works in Python, which was very helpful for this program.