Day 8: Calorie Tracker
Today, I decided to do a useful program for people to make use of. I used to always track my calories when I was dieting and since I recently learned about 'Dictionaries' in Python, I decided to make a calorie tracker program. This is just a simple calorie tracker as I am still a beginner but I'm really proud of what I coded today. I am also reading a new book called 'Grit' which inspired me to sit down early to code and to finish what I started. I had a hard time figuring out how to add the total calories but did it in the end! Ultimately, from this experience, I learned that trial and error is very helpful when it comes to coding.