Day 9: Stroop Test
I did a simpler version of this 3 weeks ago, when I first learned about parameters in Python. Since I have a bit more experience with coding since the first time I did this, I made the program longer. This is a program called the 'Stroop effect'. In psychology, the Stroop effect is a demonstration of cognitive interference where a delay in the reaction time of a task occurs due to a mismatch in stimuli. This program aims to test that hypothesis. Furthermore, I've been interested in making programs that would serve a greater purpose rather than just games. This is in Psychology, one of my favorite classes in school, and I feel that this program would help other studies that would replicate the original 'Stroop effect' study.