Day 11: Guess The Number
This is a simple program that tells the user to guess a randomized integer. I honestly didn't have time today to learn new concepts and make a unique program. So, I ended up making this program which I had a hard time with when I first started coding a month ago that I just gave up and didn't finish it. I was looking through my PyCharm projects, looking for inspiration, and finally decided that I will continue what I started a month ago. Ultimately, I learned that it really does take time for the concepts of coding to sink in and, eventually, I will come to a point where I would find what is hard for me now to be easy. I just have to trust the process.