Day 14: Email Database
This is a program that would allow the user to type in other's names and emails that will be saved in a database. After adding the names and emails, the program will print the database. It also allows the user to search for names in the program, making it easier to look for another person's email information. Today, I just learned more about lists and decided to make a program that would be useful for people in all careers.