Day 15: Calendar
This is a program that would print a calendar for the year entered by the user. I explored more on math functions today that would be helpful for making a calendar program, this included determining whether a year is a leap year or not, and when a month would start and end. To extend this program in the future, I would want to make it a GUI, but as for today, this would do.