Day 17: Fire Highlighter
This program highlights fires in an image by identifying pixels whose red intensity stronger than the rest of the threshold of the other pixels in the image. Those "sufficiently red" pixels are then highlighted in the image and the rest of the image is turned grey, by setting the pixels red, green, and blue values all to be the same average value. The image of this program can be found in the 'images' folder.