Day 18: Random Music
Music is a great way to shift a person's mood. Every since quarantine here in the Philippines started, I found myself relying on music for a mood shift. Whenever I feel down, all I needed was a good song that would instantly make me happy. The links in this program are some of my favorite songs ever. It consists of rap songs, pop songs, and a 70's song. All of these never fail to put a smile on my face, and I start feeling grateful. 

This program follows the format of the 'Motivate Me' program I made a few days ago. I decided to make a program that made use of TKinter to make the user press a button and 1 of 10 music links would come out in the terminal for them to listen to. I soon have to learn how to print the links in the GUI itself and not the terminal. However, though kind of repetitive, I'm satisfied with the program I made today because it's my first program that involves music.