Day 20: Warhol Filter
This was a project I did a few weeks ago as an assignment from Stanford's Code in Place. Today, I revised it and made a few changes. I remember having a hard time figuring out how to create a new image that is 6 times as big as the original image and ended up taking up a long time to do just that. The next challenge was to apply different filters throughout the images, which I didn't get on the spot but I really wanted to finish it so I ended up doing so. This is a program that generates the famous Andy Warhol effect based on the original image. 

TAKE NOTE: The 'PATCH_NAME' function can be edited as this program won't work if this specific image is not on your computer.