Day 21: Tough Love
We all need a little bit of tough love sometimes. This is a program that asks for the user's bad habit and would randomize a reply that would discourage the user to keep doing the habit. 

I was personally having an off day today, I didn't really feel like getting up in the morning and coding. I rested a bit and decided to get some website inspiration from http://jenniferdewalt.com that would give me some ideas for a program today. I found a basic one that I could do, named 'tough love'. So, for today's code, I just went back to basics.