Day 22: Image Reflection
I decided to experiment on images in Python today. This program takes an image, generates a new image with twice the height. The top half of the image is the same as the original. The bottom half is the mirror reflection of the top half. In other words, this program just makes the referred image (which you have to edit in the code) reflected vertically.