Day 23: Green Screen
This program shows an example of "greenscreening". This is where we replace the pixels of a certain color intensity in a particular channel (here, we use red) with the pixels from another image. I've always wondered how greenscreening worked, especially since I see it all the time now in zoom meetings. Hence, why I decided to make a program that would greenscreen 2 images.