Day 24: Memorizer (a memorizing tool)
Today I made a memorizing tool that could help students since I wanted to make use of dictionaries in Python. This program allows the user to enter a list of concepts and their respective definitions into a dictionary, shows the dictionary, lets the user look names up, and trains the user to memorize these concepts. I use Quizlet a lot when I study for exams and this is my (beginner) attempt on making my own study program that others could benefit from.