Day 25: Word Guess
I made a word guessing program today. This was part of the standard final project ideas for Code in Place. I just wanted to do the BrickBreaker initially as I had a hard time doing this. However, I finally completed it today! I had a hard time debugging this code through trial and error. Ultimately, I learned that trial and error are not as effective and productive if I have no idea what I'm doing. My goal now is to just familiarize and really know the concepts so it would be easier for me to code and debug a program.