Day 27: Trivia Game
This is a program that would test the user's trivia knowledge. It composes 25 questions from https://thoughtcatalog.com/samantha-newman/2020/03/250-best-trivia-questions/. The user would have to press the 'play' button every time s/he wants another question. Sometimes, the program would repeat a question as I made use of the random command available in Python. This program also an 'exit' button, for when the user wants to quit playing. Furthermore, it also checks the answer of the user.

It took me hours to complete this program as I only did it with my (beginner) knowledge. I am pretty sure a more experienced programmer can write this program faster and in less lines of code than what I did here. Nevertheless, I'm still happy I'm completed this exactly how I wanted it to look. I learned a lot about programming in a GUI. I just learned today how to check user's inputs using a GUI and will use these in my future programs. I really took all the time I had today to finish this code and I'm proud of myself for completing it. I still do have a lot to learn, but this is what I made with the knowledge that I have, and I am happy! I still do need to learn about how to shorten programs such as this and actually a lot more.