Day 6: Count Words
In this program, the user will input characters and will be given back how many words their input has. I decided to do this because counting words for an essay is essential for a student like me. I actually didn't expect this very short code as a result. I learned that the helpful programs don't have to be long and complicated.