Day 28: Compute Average
This is a program that reads score values from a user and computes their average. I got the idea to make this program as I saw my brother's report cards. Computing the average of numbers are commonly used and creating a program like so would be useful.

This is one of the two programs I'm doing today. I wasn't actually planning to create this and started on another program. However, I will probably finish my original program tomorrow as it's going to be long and will take me 2 days to complete. Regardless, I still want to create a finished program dedicated to this day, hence why I made this 'compute average' program.