Day 29: Workout Interval Timer
This is a Workout Interval Timer program. This is perfect for interval workouts, like HIIT, wherein the user needs to workout hard for short periods of time. This is customizable, which would fit different user's workout times. They would need to input their seconds of work, seconds of rest in between exercises and sets, and as well as the number of sets. 

My sister suggested this to me as she workouts using YouTube but she doesn't like the music being used. So, I created this program which would help her customize her timing and let her play her own background music using other apps.