Day 32: Paint
This is an interactive program that the user can use to draw or paint. I honestly got other codes on the internet as a reference for this one. I also customized it to my own liking. After this code, I can't say I'm comfortable with GUIs but through constant learning and practicing, I hope I can reach that point.