Day 33 - Moment of Calm
This is a program that the user can use to relax. The music that is used here is included in the 'music' file, the picture used is in the 'images' file as well. To recreate this, the user should have a PyCharm program with this python file, a folder named 'music' that includes 2mins.mp3/5mins.mp3/10mins.mp3/15mins.mp3/20mins.mp3 (the exact files can be seen in this repository), a folder named 'images' with 'calm.jpg' (also found in the repository). 

I have been meditating for about 2 months now. I found that it really helps me spend alone time with myself, meditating about my values, and thinking about my goals. It's one of my favorite parts of the day. I think everyone should try it at least once in their life, and see the benefits that it gives. I've also been reading tons of self-help books and a lot of them suggest meditating. Overall, I'm really happy with the results of this. I didn't stress over making it, like how I did with the other GUIs. I didn't also use reference codes but I did spend quite a lot of time with just figuring out how to play music using python. This is my first program with music and I am really happy about the result.