Day 34: Habit Tracker
This is a habit tracker program. Much like a checklist, this makes the user track his/ her habits on a weekly basis. 

This program took me literally the whole day to complete. This is also really long, but I just stuck to what I know. I learned that I still have a lot to learn when it comes to GUI programming and coding in general.