Day 36: Random Number Generator (TKinter)
For today's program, I coded a random number generator using a GUI. I was only able to do this in the terminal before, but since I've been creating GUI programs, I decided to transform it into a GUI using Tkinter. The user inputs how many numbers they need, as well as the minimum and the maximum number of the set of numbers.