Day 39: Sepia Filter
This program will make any referenced picture in the 'images' folder (that is inside the same folder as the code, or else it won't work) have a sepia filter. Sepia filter is one of the most popular tools for editing images. It makes the image appear to be brown and have a vintage-like feel.