Day 40: Unblur Movie Poster (Game)
For today's program, I created a game in which people are asked to identify the movie that corresponds with a presented movie poster. Initially, images are really blurred, but as people unsuccessfully guess, the game begins to get easier as it starts getting unblurred. People earn more points if they are able to guess the movie earlier when the image is blurrier.