Day 41: Terminal Calculator
For today's program, I made a terminal calculator. My goal for tomorrow is to convert this into a GUI. This was one of the first few skills I learned in Python, and I wanted to do a simple program to refresh my memory on it to help me prep for the GUI version tomorrow. 
The user will be asked if s/he wants to add, subtract, multiply or divide two float numbers. From there, the program will do exactly that.