Day 42: Filter V.1
For today, I made a program that would allow the user to edit their images. The options for editing include blur, contour, detail, emboss, find edges, smoothen, and sharpen. The original image will be the constant that can be seen in the GUI as PIL only allows the displayed picture to be an actual existing picture that is uploaded on the computer. Instead, I just used the '.show()' command so the users can upload their edited images. When the user clicks a button, a newly edited image will pop up on their screens. This is only version 1 as I plan to make better versions soon.