Day 45: Uni Expense Calculator
This program is useful for incoming university students as it allows them to calculate any costs that they will have, in order to effectively plan their finances for college life. This is specifically applicable to me as an incoming Uni student, as I recently got into my university of choice; this tool will allow me to keep organized throughout my life in university. A pie chart will be given to the user after inputting all the data needed as a visual guide.