Day 48: Open an Envelope
Today's program is a GUI that would tell the user to open an envelope. Once an envelope is pressed, a new window will pop up. The window contains a supportive/ advice-giving message that aims to inspire the user. 

I got this idea off of Facebook and wanted to recreate it in code the minute I saw it. It took me a few hours to finish this program but I definitely did learn a lot (opening a new window, matching colors). Overall, I'm really satisfied with my outcome.