Day 50: Monthly Expense Calculator
Today's program is a 'Monthly Expense Calculator'. A series of questions will be asked that will take in the user's monthly expenses. This tool aims to give the user the most accurate estimate and hope to keep them financially sound by giving you a visual pie chart in the end.