Day 51: Analog Clock
Today, I coded an analog clock using Turtle. This analog clock accurately shows the specific time the user is in (no matter the timezone) as it takes the current time of the user's computer.

My goal for today was just to use Turtle as I have never used it. Thankfully, I was guided with YouTube (https://www.youtube.com/watch?v=3i6iuOeEVLo) and StackOverflow to figure it out and code this. I will definitely code more programs on my own using Turtle now that I know the gist of it.