Day 52: Snake Game
This program is the classic 'Snake' game. I was searching for projects that would make me learn more about using Turtle, and this one caught my eye. I only knew how to make games using TKinter so I thought it would be great to know how to make games using Turtle. With the help of the internet and YouTube, I was able to create this 'Snake' program. It took me a while to get the hang of it, encountered some problems and felt like I'd work on this forever, but I eventually finished it and I'm really happy about how it turned out!