Day 53: Turtle Race
Today's program is a 'Turtle Race' game. This is a game with racing turtles. To make it interactive, the user can make bets with his/her friends and see which turtle finishes first. For example, the bets could be "whoever's turtle finishes first will be the first one to use the PS4", etc. 

It was fun to see the concepts I learned while only using the terminal to be applied in a game such as this. I also learned more about programming with turtle! As of now, I did use some reference code from the internet as I am still learning. In the future, my goal is to make a program on my own using turtle without a reference code.