Day 54: Turtle-Snake Game
Today's program is a combination of the last 2 programs that I made -- the turtle race and snake game. This is just like the classic snake game, however, instead of the snake, I used the turtle shape. Whenever food is eaten, an additional turtle appears behind the trail of existing turtles. I also used the pendown() function whenever the user decides to move so the path of the turtles is seen.