Day 55: Falling Ticks
Today's project was inspired by my dogs because they always get ticks. This objective of this game is to not let the dog get ticks. Instead, the dog should catch the dog food that is falling from the sky. The user gets 10 scores every time the dog gets to eat one dog food (the brown circle). If the dog gets a tick, the lives get decreased. The user only gets 3 lives. This program is similar to the 'avoid the lines' game I made days ago. The difference is that I made this using Turtle instead of TKinter. Furthermore, this is also more graphical with the dog acting as the user and with stuff to eat rather than just to avoid.