Day 57: Space Invaders
This is a simple version of the classic 'Space Invaders' game. The user's goal is to get as many invaders (the green object) as s/he can. The left and right buttons are used for changing the direction the player wants to face, and the up button is for increasing the speed.