Day 58: Social Distance Game
This is a program that prompts the user to practice social distancing by avoiding other people. The left and right buttons are used for controlling the direction the user wants to go to. The up button is for increasing the speed, the down button is for decreasing the speed. Once the user fails to avoid another "person", the score decreases. This game hopefully tells the user to be mindful of his/ her real-life interactions as COVID-19 spreads really fast.