Day 59: Olympic Rings
This is a simple program that draws the famous Olympic rings. My objective for today was just to learn more about how to code shapes and logos using Turtle. I settled on the Olympic sign as I wanted to learn how to properly position these shapes, and I eventually learned how to by coding this.