Day 60: Eight Ball V.2 (TKinter)
This is an '8 Ball' program that allows the user to ask a question, which the program will give a random answer to. I made the simple terminal version of this program when I was just starting out to learn how to code. As my skills grew, I decided to make a GUI version of it.