Day 62: Tic Tac Toe V2
This is a Tic Tac Toe game for 2 players. I used a GUI called TKinter, an upgrade to what I did yesterday which was just in the terminal. I made this with the help of https://www.youtube.com/watch?v=RuYg_D_ppYk. Although I had a reference code for this program, I do understand everything that goes on in this code.