Day 63: Click Challenge
This is a program that would test how many clicks the user can get per second. They have to click start and they'll have 60 seconds to click the 'click' button many times as they can.

I made this program on my own which I'm really happy about after days of making programs with references. This is also my first time using progress bars and found that it is easy to make use of. Overall, I'm really happy about this program!