Day 65: What's my IP Address?
This is a program that would get the user's IP address and show it to them. I got this idea from https://jenniferdewalt.com/whats_my_ip/page, when I was looking for a simple program to make. This only took a few minutes but also taught me how to position widgets in the GUI screen.