Day 66: Hangman V1 (Terminal)
This is the classic 'hangman' game. The goal for today is just to put it into the terminal first, and I will make a GUI version tomorrow. Overall, I had a really fun time trying to create this program. It took me back to Python basics which forced me to use everything I learned to make this cool program that a lot of users can enjoy.