Day 67: Hangman V.2 (TKinter)
This is the TKinter version of the Hangman game I made yesterday. This was honestly really hard and I was stuck for so long. I had a reference code for this, which is https://www.youtube.com/watch?v=esHcFbNlNqQ and https://github.com/jddelia/hangman as I had no idea how to continue with it. Honestly, I didn't fully create this code because I got frustrated so I went to the references for help. However, I did learn a lot and I'm happy with how it turned out. I did get sad when I was stuck so my goal for the remaining days is just to practice deliberately and take notes of how to solve problems in my mind so it's easier for me when I encounter problems.