Day 68: Google Dinosaur Bot
Today's program is a google dinosaur bot. It basically automates the 'No Internet' dinosaur game we all know in Google. It's my first time creating a bot and I accomplished this with the help of https://becominghuman.ai/google-chrome-dinosaur-game-python-bot-b698723e86e. Overall, I'm really happy about how it turned out. Although not perfect, it's a good introduction to bot programming!