Day 66: Heads Up V.2 (TKinter)
This is a newly improved GUI version of the Heads Up program I made during my first week of this project. This is a multiplayer game. One person would be the "it" player and s/he's not allowed to look at the screen. The other players' goal is to either describe or act out the word shown on the screen for the "it" player to guess correctly. 
I am really happy about how this turned out. I didn't use any reference code and just coded using my memory, which goes to show that I have been learning a lot and improving.