Day 72: Discount Calculator
This is a program that allows the user to compute the actual price of a product that has a discount. I made a 'Sales Tax' program 2 days ago and thought it was as useful to make this 'discount' program as well. This is specifically useful these days as a lot of shops are offering big discounts because of the pandemic.