Day 73: Alarm Clock
This 'Alarm Clock' program makes use of both the TKinter GUI and the Terminal. The user will be given a digital clock as GUI, however, the alarm time will be set using the terminal. Once the alarm is set, the user will have to wait until a YouTube browser pops up, which is the sound of the alarm.