Day 78: Fake Chat
This program is a fake chat application. This will make it look like the user is having a conversation with someone else. The time and date of when the text has been sent is also part of the speech bubble. This is especially useful for creating stories that would need fake conversations.