Day 79: Guess the Number V.2 (TKinter)
This is the TKinter version of the 'Guess the Number' program I made months before when I first started learning how to code. This program is just for numbers 1-10 and my goal is to make a nicer GUI program of this game but with numbers 1-100. I did this code with as I'm not really confident with https://levelup.gitconnected.com/learn-python-by-building-a-gui-guessing-game-with-tkinter-9f82291db6 creating this on my own. However, I did learn a lot from today and I feel confident to make the even more improved version of this program in a few days.