Day 80: Random Password Generator
This is a program that would generate a random password of 8 characters. Each time the program is run, a new password will be generated randomly. The passwords generated will be 8 characters long and will have to include the following characters in any order: 3 uppercase letters from A to Z, 3 lowercase letters from a to z, 2 digits from 0 to 9, 2 punctuation signs such as !, ?, “, # etc. To solve this challenge, I needed to know about the ASCII code which really helped to make this program shorter and more effective.