Day 82: Mad Libs Game
Mad Libs is a phrasal template word game where one player prompts others for a list of words to substitute for blanks in a story before reading aloud. This program consists of 4 mad libs templates for the user to fill in: vacation, hall pass, a valentine's day card, and a party invitation. 
It was so much fun coding this program. This is a really long code and I'm pretty sure there's a shorter way to make this program. Nevertheless, I'm really happy and proud of what I did today!