Day 83: Caesar Cipher
This is a simple python Caesar Cipher encryptor and decryptor. A Caesar cipher, also known as a shift cipher, is one of the simplest and most widely known encryption techniques. It is a type of substitution cipher in which each letter in the plaintext is replaced by a letter some fixed number of positions down the alphabet.
This is just done in the terminal as I wanted to focus on my encryption and decryption skills for today.