Day 84: Thank You Card for Frontliners
This program is a thank you card created with the turtle module. During this time of uncertainty because of the COVID-19 pandemic, the rainbow serves as a symbol to thank the medical staff for working in hospitals as well as all the frontliners who are working to provide essential services to their community. In the Philippines, these essential workers are called "frontliners". 
I just wanted to practice giving gratitude to others today by making a program like this. I'm so happy about how it turned out!