Day 86: Heads or Tails?
This is a coinflip game program made using TKinter. The user just has to press the button and s/he will either get heads or tails. The process I did here is similar to the one I did for Eightball. It's a simple code and my next idea for a program is to make a game using this 'Heads or Tails' program instead of just pressing the button.