Day 87: House (Turtle)
This is a program that would draw a house using the turtle module. This is not an interactive program as I just wanted to focus on positioning and learning how to draw the shapes today. The idea was taken from https://www.101computing.net/python-turtle-my-house/. And although I didn't fully create this code, I did learn a lot today!