Day 89: Barcode Generator
On a barcode the pattern made using black and white lines is a binary code! Black line = 1 and white line = 0. So to produce our own simplified bar codes all we need to know is convert each digit of the actual number into binary (using 4 digit binary codes, aka a nibble). So, for today, I made a program that would generate a barcode every time the user runs it. 