Day 90: Name the Color
Designers often use specific colors to design eye-catchy user interfaces. To do so they use color codes that are available on the RGB color palette. For today's program, I wrote a program that asks the user to type the name of the color and the program will return its matching color code.