Day 91: How Eco-friendly are You?
Throughout your everyday life, whatever you do, you are having an impact on the environment. For today's program, as I've never coded a quiz, I decided to design a quiz that people can take to find out how eco-friendly they are. The quiz will consist of eight questions and will be used to give the end-user a score but the user can edit this code as much as s/he'd like. This idea was taken from https://www.101computing.net/how-eco-friendly-are-you/ but I decided to make it using TKinter. I have to admit, I need more work on my designing skills, which I will be focusing on in the next programs to improve it.