Day 92: 3D Bounce
For today's program, I used Glowscript to create a 3D animation of a ball bouncing surrounded by walls. This is my first ever 3D program and I'm glad I got the hang of it fast as it is somewhat easy to understand. I had the hardest time figuring out the positioning of the walls but with a little browsing, I finally got to figure it out. Overall, I'm really happy about how this turned out.