Day 93: Honeycomb Challenge
A honeycomb is a structure of hexagonal cavities (cells of wax), made by bees to store honey and eggs. I used a set of iterative algorithms to draw a honeycomb pattern. I take a long time trying to figure out how to sort out the hexagons in order to make it look like a honeycomb. I figured it out in the end but this experience taught me that Math and logic are really important when it comes to programs such as these when they require a precise order. With that being said, I will aspire to fill in my shortcomings by practicing Math when I can so that I become proficient with making programs such as these.