Day 94: Enigma Encoder
This is an enigma encoder, which was heavily used by the German army during World War II to encrypt radio signals. Once a key is pressed on the keyboard, it goes through the plugboard which provides the first stage of the encryption process, which is based on the principles of a substitution cipher.
This is just done in the terminal as I wanted to become more proficient with my encryption and decryption skills. Thus, for today, I just learned more about more complex encoders by watching tutorials and looking at websites. This site: https://www.101computing.net/enigma/enigma-M3.html helped me the most.