Day 95: Lunar Craters
The Moon’s surface is covered with thousands of craters. For today's program, I used Python Turtle module to create a drawing of the Moon with a random selection of craters. This challenge consists of randomly positioning craters of various sizes on the surface of the Moon, the challenge is to ensure that the craters fit within the disc representing the Moon. To do this, I used a set of random Cartesian coordinates (x,y). Before using these coordinates to draw a crater we calculate the distance between the center of the disc representing the Moon and the randomly generated (x,y) coordinates. To do so I used the Pythagoras’ Theorem. If the distance is lower than (Radius of the Moon – Radius of Crater) then we keep these coordinates to draw our crater, otherwise, we generate a new set of coordinates.