Day 96: Temperature Converter (Terminal)
Ever since the start of quarantine, to ease my boredom, I have been baking and cooking a lot using the recipes I find online. I wanted to create a program that would make it faster to convert units of measurement. For today's program, I made a temperature converter that would convert between the most commonly used units to measure: Degree Fahrenheit (°F) and Degree Celsius (°C). This program will ask the user whether they want to convert from Celsius to Fahrenheit or from Fahrenheit to Celsius, ask the user to enter a temperature in the correct unit, and apply the conversion formula and display the temperature in the new unit. This idea was taken from https://www.101computing.net/flowchart-to-python-code-temperature-converter/ which guided me throughout the process.