Day 98: Radians to Degrees Conversions
The radian is a unit of measure for angles used mainly in trigonometry. It is used as an alternative to degrees. In degrees a full circle is 360 degrees, in radians a full circle is 2π radians. It is easy to convert an angle from degrees to radians and vice-versa using the following formulas: radians = degrees x π / 180 and degrees = radians x 180 / π. For today, I made a program to convert an angle from degrees to radians or from radians to degrees based on the user’s choice.