Day 99: 3D Chicken
For today's program, I used GlowScript to create a 3D of Crossy Road's chicken. To achieve this, I used a 2D array data structure to represent a graphic. In Python, 2D arrays are implemented using a list of lists. In this code, I declared a 2D array called pixels to store the value of each pixel used to represent one of the chicken. Using 2 nested loops I then looped through each row and each column of this array to retrieve the value of each pixel in order to create the cubes needed to create a 3D model of a chicken. Finally, I then created a compound object to join these cubes together in a single object.