Day 107: Christmas Tree
Christmas is coming!! In the Philippines, people get into the holiday spirit on the first day of the 'BER' months (September 1). This is when the fun starts and when the countdown to the long Christmas in the Philippines begins! For today's program, I created Christmas Cards using Python's turtle module as a way of getting into the Christmas spirit myself.