Day 108: 3D Traffic Lights
For today, I made another 3D program. This program consists of two traffic lights used at a crossing. A traffic light control system is an example of an embedded system. An embedded system is a computer system with a dedicated function within a larger mechanical or electrical system, often with real-time computing constraints. For this program, I created algorithms used by these control systems that focus on the traffic lights system.