Day 114: Weather App
This is a Weather application that allows the user to type in any city, and then the program returns the real-time weather data of the city. My objective for today was just to learn about connecting GUIs to APIs and making a usable application with it. I learned how to do this program from: https://www.youtube.com/watch?v=D8-snVfekto. With this new knowledge, I plan to make more useful applications that use real-time APIs such as this!