Day 116: Color Game
In this game, the player has to enter the colors of the words that appear on the screen within 30 seconds. The interface will display the name of different colors in different colors. The Player has to identify the color and enter the correct color name to win the game. I did a terminal version of this Stroop test/ color game program on the 9th day of this project. Like back then, this program aims to test out the hypothesis. This program aims to help others replicate the Stroop effect study digitally.