Day 117: Age Calculator
I honestly didn't have as much time to create an experimental/ big project due to personal reasons. Even though, I created this program that was sufficient for the time I had. In this age calculator app, the user can type in their date of birth, and the app will calculate and display their age.