Day 119: Student Management System
This repository contains a Student Management System. This program makes use of Tkinter and sqlite3. It asks the user to enter their Name, University, Course, Email & Student Number. This follows the format of the project I made yesterday. I made this program as I believe it would be useful for teachers especially in this time of online distant learning.