Day 120: Bookshelf
This is a 'Bookshelf' program that I made using TKinter and sqlite3. You can use Bookshelf as a storeroom of the books which you have read. The Bookshelf stores the information about the Title and Author of the book and the Date the user finished reading. Users can: view all entries, add a new entry, search, update and delete an entry, and close the program via the corresponding buttons. The design of the GUI follows the format of my last 2 previous programs.