Day 122: Clear Paint
For today, I made a simple drawing program. I wanted to be able to doodle with my computer on a semi-transparent surface while a video, a picture, or something is in the background. This is similar to the 'Paint' program I made before but this one has a clear screen and it has an different design overall.