Day 123: Cafe Management System
I went to a Cafe for the first time in a really long time today. So, for today's program, I made a 'Cafe Management System' that is similar to the Restaurant version of what I made days ago but this one has an added 'receipt' function. I found it really fun creating this code and imagining myself as someone who owns a cafe; designing a menu and figuring out the prices. This code is also helpful for those businesses that would like to make use of this.