Day 125: Mine Sweeper
For today, I wanted to create a game as I haven't in a long time. This is the classic 'Mine Sweeper' game. The objective of the game is to clear a rectangular board containing hidden "mines" or bombs without detonating any of them, with help from clues about the number of neighboring mines in each field. (PS if you want to recreate this program and want the pictures, just contact me)