Day 126: Package Tracker
A lot of people have been ordering online these days because of the pandemic. With that in mind, I wanted to create a program that would help people track their packages. I used my past experience with creating the 'Weather App' for making this other program that uses and API. Creating this was really new and I did learn a lot throughout the process. Also, I did test it out with some tracking numbers that I saw online to test it out but didn't show it on the video because of privacy reasons. The program does work, I just didn't get to show it in the video because of privacy reasons.