Day 130: Alphabetic Order
We're going back to basics for today's program. This program can sort the words from the user's input in alphabetic order using the 'for loop' function. This is a very basic code and it just took me a few minutes. I didn't really have time today because I've just been employed but I still found time to code this. Although easy, I believe this program can be really useful.